+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added support for `HuggingFace` embeddings models.
+
+### Changed
+
+- Improved `README.md` to include information about how to setup a test project on local GitLab.
+
 ### Fixed
 
 - PlanAndExecuteAgent: Plan prompt `<file_path>` was always empty.